#22 Moved *Components and *Settings to "Bags" subfolder.
