+++ src/NW.UnivariateForecasting/SlidingWindows/SlidingWindowManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using NW.UnivariateForecasting.Bags;
 using NW.UnivariateForecasting.Validation;
 
 namespace NW.UnivariateForecasting.SlidingWindows
