+++ src/NW.UnivariateForecastingClient/Shared/IUnivariateForecasterFactory.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using NW.UnivariateForecasting;
+using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecastingClient.Shared
 {
