+++ src/NW.UnivariateForecastingClient/Shared/IUnivariateForecastingComponentsFactory.cs
@@ -1,4 +1,4 @@
-﻿using NW.UnivariateForecasting;
+﻿using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecastingClient.Shared
 {
