+++ src/NW.UnivariateForecastingClient/Shared/UnivariateForecasterFactory.cs
@@ -1,4 +1,5 @@
 ﻿using NW.UnivariateForecasting;
+using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecastingClient.Shared
 {
