+++ tests/NW.UnivariateForecasting.UnitTests/Observations/ObservationManagerTests.cs
@@ -4,6 +4,7 @@ using NW.UnivariateForecasting.Observations;
 using NW.UnivariateForecasting.SlidingWindows;
 using NW.UnivariateForecasting.UnitTests.Utilities;
 using NUnit.Framework;
+using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecasting.UnitTests.Observations
 {
