+++ tests/NW.UnivariateForecasting.UnitTests/UnivariateForecasterTests.cs
@@ -11,6 +11,7 @@ using NW.UnivariateForecasting.Serializations;
 using NW.UnivariateForecasting.SlidingWindows;
 using NW.UnivariateForecasting.UnitTests.Utilities;
 using NUnit.Framework;
+using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecasting.UnitTests
 {
