+++ tests/NW.UnivariateForecasting.UnitTests/Utilities/ObjectMother.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Reflection;
 using NUnit.Framework;
+using NW.UnivariateForecasting.Bags;
 
 namespace NW.UnivariateForecasting.UnitTests.Utilities
 {
