+++ tests/NW.UnivariateForecastingClient.UnitTests/Shared/UnivariateForecastingFactoryTests.cs
@@ -1,4 +1,5 @@
 ﻿using NW.UnivariateForecasting;
+using NW.UnivariateForecasting.Bags;
 using NW.UnivariateForecastingClient.Shared;
 using NUnit.Framework;
 
@@ -59,5 +60,5 @@ namespace NW.UnivariateForecastingClient.UnitTests
 
 /*
     Author: numbworks@gmail.com
-    Last Update: 08.03.2023
+    Last Update: 08.02.2024
 */
